// src/app/shared/models/vehicle-inventory/vehicle.ts
export interface Vehicle {
    vin: string;
    mileage: number;
    model: string;
    year: number;
    exteriorColor: string;
    makeName: string;
    vehicleStock: string;
    stockType: string;
    modelType: string;
    price: number;
    parentId: string;
    dealerId: string;
    bodyStyle: string;
    financeCompany: string;
    pricingType: string;
}